requirements.txt
absl-py==2.3.1
accelerate==1.8.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
array_record==0.7.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==16.0.1
bcrypt==4.3.0
beartype==0.22.9
beautifulsoup4==4.13.5
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
chex==0.1.90
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
compressai==1.2.8
contourpy==1.3.2
controlnet_aux==0.0.10
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.35.1
dill==0.3.8
distrax==0.1.4
distro==1.9.0
dm-tree==0.1.9
docstring_parser==0.16
easy_dwpose==1.0.2
einops==0.8.1
etils==1.12.2
exceptiongroup==1.3.0
executing==2.2.1
fabric==3.2.2
facexlib==0.3.0
fairscale==0.4.13
filelock==3.13.1
filterpy==1.4.5
flatbuffers==25.2.10
flax==0.10.7
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.0
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
imgaug==0.4.0
immutabledict==4.2.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
invoke==2.2.0
iopath==0.1.10
ipython==8.37.0
jax==0.6.2
jax-cuda12-pjrt==0.6.2
jax-cuda12-plugin==0.6.2
jaxlib==0.6.2
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.12.0
joblib==1.5.2
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
libtmux==0.46.2
lightning-utilities==0.15.0
llvmlite==0.44.0
lmdb==1.6.2
loguru==0.7.3
lpips==0.1.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.1
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.4.99
nymeria==1.2.0
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime-gpu==1.16.2
openai==1.57.1
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opensimplex==0.4.5.1
opt_einsum==3.4.0
optax==0.2.5
optree==0.16.0
orbax==0.1.9
orbax-checkpoint==0.11.18
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.1
paramiko==3.5.1
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
playwright==1.54.0
plotly==6.2.0
portalocker==3.2.0
proglog==0.1.12
promise==2.3
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==4.21.12
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pybind11==3.0.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.2
pyiqa==0.1.10
PyNaCl==1.5.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-msssim==1.0.0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.3
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.10.0
seaborn==0.13.2
selenium==4.36.0
sentry-sdk==2.32.0
shapely==2.1.2
simple-parsing==0.1.7
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.9
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
tensorflow-probability==0.25.0
tensorflow_cpu==2.19.0
tensorstore==0.1.76
termcolor==3.1.0
threadpoolctl==3.1.0
tifffile==2025.5.10
tiktoken==0.3.3
timm==1.0.16
tinycss2==1.4.0
toml==0.10.2
tomli==2.3.0
toolz==1.0.0
torch==2.4.0+cu124
torch-dct==0.1.6
torch-geometric==2.7.0
-e git+https://github.com/LTH14/torch-fidelity.git@5746b7685c505fe3cb8ef0895c93b0d5f21f8b7d#egg=torch_fidelity
torchaudio==2.4.0+cu124
torchmetrics==1.7.4
torchvision==0.19.0+cu124
tqdm==4.67.1
traitlets==5.14.3
treescope==0.1.9
trio==0.31.0
trio-websocket==0.12.2
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
volcengine-python-sdk==1.0.111
wadler_lindig==0.1.7
wandb==0.21.0
wcwidth==0.2.14
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0